pyproject.toml
[project]
name = "results-optimization"
version = "0.1.0"
description = ""
authors = [
    {name = "Julius Irion",email = "juliusirion@gmx.de"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core (>=1.3.2,<2.0.0)",
    "hydra-optuna-sweeper (>=1.2.0,<2.0.0)",
    "hydra-joblib-launcher (>=1.2.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
